+++ src/hocs/with-video/with-video.e2e.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {mount, configure} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import withActivePlayer from './with-active-player';
+import withVideo from './with-video';
 import {func} from 'prop-types';
 
 configure({
@@ -23,7 +23,7 @@ MockComponent.propTypes = {
 
 
 it(`withActiveVideo should start playing video after timer onMouseEnter`, () => {
-  const MockComponentWrapped = withActivePlayer(MockComponent);
+  const MockComponentWrapped = withVideo(MockComponent);
   const wrapper = mount(
       <MockComponentWrapped
         onMouseEnter={() => {}}
@@ -43,7 +43,7 @@ it(`withActiveVideo should start playing video after timer onMouseEnter`, () => 
 });
 
 it(`withActiveVideo should clear timer onMouseLeave`, () => {
-  const MockComponentWrapped = withActivePlayer(MockComponent);
+  const MockComponentWrapped = withVideo(MockComponent);
   const wrapper = mount(
       <MockComponentWrapped
         onMouseEnter={() => {}}
@@ -59,7 +59,7 @@ it(`withActiveVideo should clear timer onMouseLeave`, () => {
 });
 
 it(`withActiveVideo should clear timer on componentWillUnmount`, () => {
-  const MockComponentWrapped = withActivePlayer(MockComponent);
+  const MockComponentWrapped = withVideo(MockComponent);
   const wrapper = mount(
       <MockComponentWrapped
         onMouseEnter={() => {}}
@@ -72,3 +72,62 @@ it(`withActiveVideo should clear timer on componentWillUnmount`, () => {
 
   expect(clearTimeout).toHaveBeenCalledTimes(1);
 });
+
+/*
+VideoPlayer tests
+
+import React from 'react';
+import {configure, mount} from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+import VideoPlayer from './video-player';
+
+configure({
+  adapter: new Adapter(),
+});
+
+
+it(`VideoPlayer should play`, () => {
+  const isPlaying = true;
+  const player = mount(
+      <VideoPlayer
+        src={``}
+        poster={``}
+        muted={true}
+        isPlaying={isPlaying}
+        onMouseEnter={() => {}}
+        onMouseLeave={() => {}}
+      />
+  );
+
+  HTMLMediaElement.prototype.play = jest.fn();
+
+  player.instance().componentDidUpdate();
+
+  const isPlay = player.instance().props.isPlaying;
+  expect(isPlay).toBe(true);
+  expect(HTMLMediaElement.prototype.play).toHaveBeenCalledTimes(1);
+});
+
+it(`VideoPlayer should stop playing and start loading`, () => {
+  const isPlaying = false;
+  const player = mount(
+      <VideoPlayer
+        src={``}
+        poster={``}
+        muted={true}
+        isPlaying={isPlaying}
+        onMouseEnter={() => {}}
+        onMouseLeave={() => {}}
+      />
+  );
+
+  HTMLMediaElement.prototype.load = jest.fn();
+
+  player.instance().componentDidUpdate();
+
+  const isPlay = player.instance().props.isPlaying;
+  expect(isPlay).toBe(false);
+  expect(HTMLMediaElement.prototype.load).toHaveBeenCalledTimes(1);
+});
+
+*/
