+++ src/hocs/with-video/with-video.js
@@ -1,10 +1,12 @@
-import React, {PureComponent} from 'react';
+import React, {PureComponent, createRef} from 'react';
+import {string} from 'prop-types';
 import {DELAY} from '../../utils/consts';
 
 const withActivePlayer = (Component) => {
   class Wrapper extends PureComponent {
     constructor(props) {
       super(props);
+      this._videoRef = createRef();
       this.state = {
         isPlaying: false,
       };
@@ -14,7 +16,37 @@ const withActivePlayer = (Component) => {
       this._handleMovieCardMouseLeave = this._handleMovieCardMouseLeave.bind(this);
     }
 
+    componentDidMount() {
+      const {src, poster} = this.props;
+      const video = this._videoRef.current;
+      if (video) {
+        video.src = src;
+        video.poster = poster;
+        video.muted = true;
+      }
+    }
+
+    componentDidUpdate() {
+      const video = this._videoRef.current;
+      const {isPlaying} = this.state;
+      if (!video) {
+        return;
+      }
+      if (isPlaying) {
+        video.play();
+      } else {
+        video.load();
+      }
+    }
+
     componentWillUnmount() {
+      const video = this._videoRef.current;
+      if (video) {
+        video.src = ``;
+        video.poster = null;
+        video.muted = null;
+        video.onplay = null;
+      }
       clearTimeout(this._timerId);
     }
 
@@ -35,9 +67,21 @@ const withActivePlayer = (Component) => {
         isPlaying={isPlaying}
         onMouseEnter={this._handleMovieCardMouseEnter}
         onMouseLeave={this._handleMovieCardMouseLeave}
-      />;
+      >
+        <video
+          ref={this._videoRef}
+          width="280"
+          height="175"
+          preload="none"
+        />
+      </Component>;
     }
   }
+  Wrapper.propTypes = {
+    src: string.isRequired,
+    poster: string.isRequired,
+  };
+
   return Wrapper;
 };
